Add install targets to makefiles
Add install targets and export directives to the cmake config, enabling
install [1] and package discoverability [2]. This allows for simple
find_package use in downstream cmake projects, and predictable behaviour
when building azure-storage-lite with the intention of installing it.

Headers are installed to <include>/azure-storage-lite, to avoid name
clashes with other packages.

[1] cmake --target install, make install or similar
[2] find_package(azure-store-lite)